% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedSearch.R
\name{search}
\alias{search}
\title{Pubmed Search}
\usage{
search(keyword)
}
\arguments{
\item{keyword}{Character string of pubmed search}
}
\value{
Number of results from a search of pubmed database
}
\description{
Pubmed Search
}
\examples{
search("nursing[MeSH]")


}
